backend/requirements.txt
Flask==3.1.1
flask_cors==6.0.0
mysql_connector_repackaged==0.3.1
neo4j==5.28.1
openai==1.82.0
pandas==2.2.3
Pillow==11.2.1
pymilvus==2.5.8
PyMySQL==1.1.1
pypinyin==0.54.0
rapidfuzz==3.13.0
Requests==2.32.3
sentence_transformers==4.1.0
Werkzeug==3.1.3
